\name{dad_joke_print}
\alias{dad_joke_print}
\title{Print a Dad Joke to the Console}
\description{
  Fetches a random dad joke from the JokeAPI and prints it to the console.
}
\usage{
dad_joke_print()
}
\arguments{
  \item{}{} No arguments.
}
\value{
  Nothing. It prints the joke to the console.
}
\examples{
# To print a dad joke to the console
dad_joke_print()
}
