\name{progr_joke_img}
\alias{progr_joke_img}
\title{Generate and Save a Programmer Joke Image}
\description{
  Fetches a random programming joke from the JokeAPI and generates an image with the joke text.
  The image is saved as `programmer_joke.png` in the current working directory.
}
\usage{
progr_joke_img()
}
\arguments{
  \item{}{} No arguments.
}
\value{
  The file path of the saved image.
}
\examples{
# To generate and save the image with a programmer joke
progr_joke_img()
}
