\name{progr_joke_print}
\alias{progr_joke_print}
\title{Print a Programmer Joke to the Console}
\description{
  Fetches a random programming joke from the JokeAPI and prints it to the console.
}
\usage{
progr_joke_print()
}
\arguments{
  \item{}{} No arguments.
}
\value{
  Nothing. It prints the joke to the console.
}
\examples{
# To print a programmer joke to the console
progr_joke_print()
}
